Move fetcher to importer \ fetcher
