+++ src/main/java/net/sf/jabref/external/FindFullTextAction.java
@@ -26,7 +26,7 @@ import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.gui.FileListTableModel;
 import net.sf.jabref.gui.undo.UndoableFieldChange;
 import net.sf.jabref.gui.worker.AbstractWorker;
-import net.sf.jabref.logic.fulltext.FindFullText;
+import net.sf.jabref.logic.importer.FindFullText;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
