+++ src/main/java/net/sf/jabref/logic/importer/FindFullText.java
@@ -1,4 +1,21 @@
-package net.sf.jabref.logic.fulltext;
+/*
+ * Copyright (C) 2003-2016 JabRef contributors.
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
+package net.sf.jabref.logic.importer;
 
 import java.io.IOException;
 import java.net.URL;
@@ -7,6 +24,13 @@ import java.util.List;
 import java.util.Optional;
 
 import net.sf.jabref.importer.fetcher.CrossRef;
+import net.sf.jabref.logic.importer.fetcher.ACS;
+import net.sf.jabref.logic.importer.fetcher.DoiResolution;
+import net.sf.jabref.logic.importer.fetcher.GoogleScholar;
+import net.sf.jabref.logic.importer.fetcher.IEEE;
+import net.sf.jabref.logic.importer.fetcher.ScienceDirect;
+import net.sf.jabref.logic.importer.fetcher.SpringerLink;
+import net.sf.jabref.logic.importer.fetcher.ArXiv;
 import net.sf.jabref.logic.util.DOI;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
