+++ src/main/java/net/sf/jabref/logic/importer/fetcher/ArXiv.java
@@ -1,4 +1,21 @@
-package net.sf.jabref.logic.fulltext;
+/*
+ * Copyright (C) 2003-2016 JabRef contributors.
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
+package net.sf.jabref.logic.importer.fetcher;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
@@ -17,16 +34,18 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
 import net.sf.jabref.gui.help.HelpFile;
-import net.sf.jabref.importer.fetcher.IdBasedFetcher;
+import net.sf.jabref.logic.importer.IdBasedFetcher;
 import net.sf.jabref.importer.fetcher.OAI2Fetcher;
-import net.sf.jabref.importer.fetcher.SearchBasedFetcher;
+import net.sf.jabref.logic.importer.SearchBasedFetcher;
 import net.sf.jabref.logic.TypedBibEntry;
-import net.sf.jabref.logic.fetcher.FetcherException;
+import net.sf.jabref.logic.importer.FullTextFinder;
+import net.sf.jabref.logic.importer.FetcherException;
 import net.sf.jabref.logic.util.DOI;
 import net.sf.jabref.logic.util.io.XMLUtil;
 import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.model.database.BibDatabaseMode;
 import net.sf.jabref.model.entry.BibEntry;
+import net.sf.jabref.model.entry.BibtexEntryTypes;
 import net.sf.jabref.model.entry.FieldName;
 import net.sf.jabref.model.entry.ParsedFileField;
