+++ src/test/java/net/sf/jabref/logic/fulltext/FindFullTextTest.java
@@ -5,6 +5,8 @@ import java.net.URL;
 import java.util.Arrays;
 import java.util.Optional;
 
+import net.sf.jabref.logic.importer.FindFullText;
+import net.sf.jabref.logic.importer.FullTextFinder;
 import net.sf.jabref.model.entry.BibEntry;
 
 import org.junit.After;
